requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.4.1
python-dotenv==1.0.1
Werkzeug==3.1.3
SQLAlchemy==2.0.38
gunicorn==23.0.0
requests==2.32.3
psycopg[binary]==3.3.3
redis==5.2.1
qrcode[pil]==8.0
simple-websocket==1.1.0
pytest==8.3.4
